+++ src/sll/io/generate_code.c
@@ -569,6 +569,9 @@ static const sll_node_t* _generate_code_internal(const code_generation_data_t* c
 		case SLL_NODE_TYPE_ACCESS:
 			sll_file_write_char(wf,':');
 			break;
+		case SLL_NODE_TYPE_DEEP_COPY:
+			PRINT_STATIC_STRING(":!",wf);
+			break;
 		case SLL_NODE_TYPE_VAR_ACCESS:
 			{
 				sll_arg_count_t l=o->dt.ac;